requirements.txt
# DDR Report Generator — Requirements
# Install:  pip install -r requirements.txt
# Run:      streamlit run app.py

streamlit>=1.35.0       # Web UI
groq>=0.9.0             # Groq API client (Llama 3.3 70B)
python-dotenv>=1.0.0    # Loads GROQ_API_KEY from .env file
pdfplumber>=0.11.0      # PDF text extraction
pypdf>=4.3.0            # PDF image extraction (no poppler needed)
reportlab>=4.2.0        # PDF report creation
Pillow>=10.0.0          # Image processing
